fix: resolve clippy warnings and add cross-platform check script
- Fix clippy::io_other_error lints (use std::io::Error::other)
- Fix clippy::manual_div_ceil lints (use .div_ceil())
- Fix clippy::missing_transmute_annotations in DeviceWrapper
- Fix clippy::redundant_closure_call in TUI
- Run cargo fmt across entire workspace
- Add scripts/cross-check.sh for local cross-platform validation
  using cargo-zigbuild (zig as cross-compiler) to check Linux
  and Windows compilation without burning GitHub Actions minutes

+++ crates/drivewipe-gui/src/screens/confirm.rs
@@ -32,7 +32,7 @@ pub fn view<'a>(
 
     let input = text_input("Type YES", confirm_text)
         .on_input(Message::ConfirmInput)
-        .size(theme::FONT_SIZE_LG as f32);
+        .size(theme::FONT_SIZE_LG);
 
     let is_confirmed = confirm_text.trim() == "YES";
